Revert "Added has() and belongs_to() association support"
This reverts commit d6c31b00a588e95a4d221e3b45ffacb2e042bd5d.

While this behavior is consistent with Ambition, it is not consistent
with Enumerable behavior.  A block created using this API could not
be used to iterate over an Array of Resource objects, which is the
goal.  The API was:

  User.select { |u| u.articles.title == 'DataMapper' }

When it should be:

  User.select { |u| u.articles.all? { |a| a.title == 'DataMapper' } }

Sure the above is not quite as concise, but it is closer to being
indistinguishable from pure Ruby, which is the ultimate goal.

Another reason is that if the first syntax was used to iterate over an
already loaded Collection, then an exception would be thrown since there
is no #title method for the User.has(n, :articles) relationship.